+++ src/animation-clip-data.ts
@@ -1,4 +1,4 @@
-import type { EffectsObjectData } from './components';
+import { EffectsObjectData } from './effects-object-data';
 import type { FixedNumberExpression, FixedVec3Expression, FixedQuatExpression } from './number-expression';
 
 export interface AnimationClipData extends EffectsObjectData {
