+++ src/components/component-data.ts
@@ -1,0 +1,9 @@
+import { EffectsObjectData } from "../effects-object-data";
+
+export interface DataPath {
+  id: string,
+}
+
+export interface ComponentData extends EffectsObjectData {
+  item: DataPath,
+}
