+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./component-data";
+export * from "./post-process-volume-data";
