+++ src/components/post-process-volume-data.ts
@@ -1,0 +1,18 @@
+export interface PostProcessVolumeData {
+  bloomEnabled: boolean,
+  threshold: number,
+  bloomIntensity: number,
+
+  // ColorAdjustments
+  brightness: number,
+  saturation: number,
+  contrast: number,
+
+  // Vignette
+  vignetteIntensity: number,
+  vignetteSmoothness: number,
+  vignetteRoundness: number,
+
+  // ToneMapping
+  toneMappingEnabled: boolean,
+}
