+++ src/data-type.ts
@@ -1,0 +1,50 @@
+export enum DataType {
+  VFXItemData = 'VFXItemData',
+  EffectComponent = 'EffectComponent',
+  Material = 'Material',
+  Shader = 'Shader',
+  SpriteComponent = 'SpriteComponent',
+  ParticleSystem = 'ParticleSystem',
+  InteractComponent = 'InteractComponent',
+  CameraController = 'CameraController',
+  PostProcessVolume = 'PostProcessVolume',
+  Geometry = 'Geometry',
+  Texture = 'Texture',
+  AnimationClip = 'AnimationClip',
+  TextComponent = 'TextComponent',
+  BinaryAsset = 'BinaryAsset',
+
+  // Timeline
+  TrackAsset = 'TrackAsset',
+  TimelineAsset = 'TimelineAsset',
+  ObjectBindingTrack = 'ObjectBindingTrack',
+  TransformTrack = 'TransformTrack',
+  SpriteColorTrack = 'SpriteColorTrack',
+  ActivationTrack = 'ActivationTrack',
+  SubCompositionTrack = 'SubCompositionTrack',
+  FloatPropertyTrack = 'FloatPropertyTrack',
+  ColorPropertyTrack = 'ColorPropertyTrack',
+  Vector4PropertyTrack = 'Vector4PropertyTrack',
+
+  TransformPlayableAsset = 'TransformPlayableAsset',
+  SpriteColorPlayableAsset = 'SpriteColorPlayableAsset',
+  ActivationPlayableAsset = 'ActivationPlayableAsset',
+  SubCompositionPlayableAsset = 'SubCompositionPlayableAsset',
+  FloatPropertyPlayableAsset = 'FloatPropertyPlayableAsset',
+  ColorPropertyPlayableAsset = 'ColorPropertyPlayableAsset',
+
+  // FIXME: 先完成ECS的场景转换，后面移到spec中
+  MeshComponent = 'MeshComponent',
+  SkyboxComponent = 'SkyboxComponent',
+  LightComponent = 'LightComponent',
+  CameraComponent = 'CameraComponent',
+  ModelPluginComponent = 'ModelPluginComponent',
+  TreeComponent = 'TreeComponent',
+  AnimationComponent = 'AnimationComponent',
+  SpineComponent = 'SpineComponent',
+  VideoComponent = 'VideoComponent',
+  AudioComponent = 'AudioComponent',
+
+  // Non-EffectObject
+  TimelineClip = 'TimelineClip',
+}
