+++ src/effects-object-data.ts
@@ -1,0 +1,7 @@
+import { DataType } from "./data-type";
+
+export interface EffectsObjectData {
+  id: string,
+  name?: string,
+  dataType: DataType,
+}
