+++ src/effects-package-data.ts
@@ -1,0 +1,12 @@
+import { EffectsObjectData } from "./effects-object-data";
+
+export interface EffectsPackageData {
+    fileSummary: FileSummary,
+    exportObjects: EffectsObjectData[],
+  }
+  
+  export interface FileSummary {
+    guid: string,
+    assetType: string,
+  }
+
