+++ src/geometry-data.ts
@@ -1,102 +1,4 @@
-import type { Vector2Data } from './item/base-item';
-
-export enum DataType {
-  VFXItemData = 'VFXItemData',
-  EffectComponent = 'EffectComponent',
-  Material = 'Material',
-  Shader = 'Shader',
-  SpriteComponent = 'SpriteComponent',
-  ParticleSystem = 'ParticleSystem',
-  InteractComponent = 'InteractComponent',
-  CameraController = 'CameraController',
-  Geometry = 'Geometry',
-  Texture = 'Texture',
-  AnimationClip = 'AnimationClip',
-  TextComponent = 'TextComponent',
-  BinaryAsset = 'BinaryAsset',
-
-  // Timeline
-  TrackAsset = 'TrackAsset',
-  TimelineAsset = 'TimelineAsset',
-  ObjectBindingTrack = 'ObjectBindingTrack',
-  TransformTrack = 'TransformTrack',
-  SpriteColorTrack = 'SpriteColorTrack',
-  ActivationTrack = 'ActivationTrack',
-  SubCompositionTrack = 'SubCompositionTrack',
-  FloatPropertyTrack = 'FloatPropertyTrack',
-  ColorPropertyTrack = 'ColorPropertyTrack',
-  Vector4PropertyTrack = 'Vector4PropertyTrack',
-
-  TransformPlayableAsset = 'TransformPlayableAsset',
-  SpriteColorPlayableAsset = 'SpriteColorPlayableAsset',
-  ActivationPlayableAsset = 'ActivationPlayableAsset',
-  SubCompositionPlayableAsset = 'SubCompositionPlayableAsset',
-  FloatPropertyPlayableAsset = 'FloatPropertyPlayableAsset',
-  ColorPropertyPlayableAsset = 'ColorPropertyPlayableAsset',
-
-  // FIXME: 先完成ECS的场景转换，后面移到spec中
-  MeshComponent = 'MeshComponent',
-  SkyboxComponent = 'SkyboxComponent',
-  LightComponent = 'LightComponent',
-  CameraComponent = 'CameraComponent',
-  ModelPluginComponent = 'ModelPluginComponent',
-  TreeComponent = 'TreeComponent',
-  AnimationComponent = 'AnimationComponent',
-  SpineComponent = 'SpineComponent',
-  VideoComponent = 'VideoComponent',
-  AudioComponent = 'AudioComponent',
-
-  // Non-EffectObject
-  TimelineClip = 'TimelineClip',
-}
-
-export interface DataPath {
-  id: string,
-}
-
-export interface EffectsObjectData {
-  id: string,
-  name?: string,
-  dataType: DataType,
-}
-
-export interface ComponentData extends EffectsObjectData {
-  item: DataPath,
-}
-
-export interface ColorData {
-  r: number,
-  g: number,
-  b: number,
-  a: number,
-}
-
-export interface Vector4Data {
-  x: number,
-  y: number,
-  z: number,
-  w: number,
-}
-
-export interface MaterialTextureProperty {
-  texture: DataPath,
-  offset?: Vector2Data,
-  scale?: Vector2Data,
-}
-
-export interface MaterialData extends EffectsObjectData {
-  shader: DataPath,
-  blending?: boolean,
-  zWrite?: boolean,
-  zTest?: boolean,
-  stringTags: Record<string, string>,
-  macros: string[],
-  ints: Record<string, number>,
-  floats: Record<string, number>,
-  vector4s: Record<string, Vector4Data>,
-  colors: Record<string, ColorData>,
-  textures: Record<string, MaterialTextureProperty>,
-}
+import { EffectsObjectData } from "./effects-object-data";
 
 export interface GeometryData extends EffectsObjectData {
   vertexData: VertexData,
@@ -231,20 +133,4 @@ export enum VertexBufferSemantic {
   TangentBS1 = 'TANGENT_BS1',
   TangentBS2 = 'TANGENT_BS2',
   TangentBS3 = 'TANGENT_BS3',
-}
-
-export interface ShaderData extends EffectsObjectData {
-  vertex: string,
-  fragment: string,
-  properties?: string,
-}
-
-export interface EffectsPackageData {
-  fileSummary: FileSummary,
-  exportObjects: EffectsObjectData[],
-}
-
-export interface FileSummary {
-  guid: string,
-  assetType: string,
 }
