+++ src/index.ts
@@ -31,3 +31,9 @@ export * from './timeline';
 export * from './assets';
 export * from './render-settings';
 export * from './curve-data';
+export * from './data-type';
+export * from './effects-object-data';
+export * from './material-data';
+export * from './geometry-data';
+export * from './shader-data';
+export * from './effects-package-data';
