+++ src/item/spine-item.ts
@@ -1,5 +1,6 @@
 import type { BinaryPointer } from '../binary';
-import type { ComponentData, DataPath, DataType } from '../components';
+import type { ComponentData, DataPath } from '../components';
+import { DataType } from '../data-type';
 import type {
   SizeOverLifetime, RotationOverLifetime, PositionOverLifetime, ColorOverLifetime,
   RendererOptions, ItemType,
