+++ src/math/color-data.ts
@@ -1,0 +1,7 @@
+
+export interface ColorData {
+    r: number,
+    g: number,
+    b: number,
+    a: number,
+  }
