+++ src/math/index.ts
@@ -1,0 +1,2 @@
+export * from './color-data';
+export * from './vector-data';
