+++ src/math/vector-data.ts
@@ -1,0 +1,18 @@
+
+export interface Vector4Data {
+  x: number,
+  y: number,
+  z: number,
+  w: number,
+}
+
+export interface Vector3Data {
+  x: number,
+  y: number,
+  z: number,
+}
+
+export interface Vector2Data {
+  x: number,
+  y: number,
+}
