+++ src/shader-data.ts
@@ -1,0 +1,7 @@
+import { EffectsObjectData } from "./effects-object-data";
+
+export interface ShaderData extends EffectsObjectData {
+  vertex: string,
+  fragment: string,
+  properties?: string,
+}
