+++ src/timeline/playables/sprite-color-playable-asset-data.ts
@@ -1,6 +1,7 @@
-import type { DataType, EffectsObjectData } from '../../components';
 import type { RGBAColorValue } from '../../number-expression';
 import type { ColorOverLifetime } from '../../type';
+import { DataType } from '../../data-type';
+import { EffectsObjectData } from '../../effects-object-data';
 
 export interface SpriteColorPlayableAssetData extends EffectsObjectData {
   dataType: DataType.SpriteColorPlayableAsset,
