+++ src/timeline/timeline-asset-data.ts
@@ -1,4 +1,6 @@
-import type { EffectsObjectData, DataPath, DataType } from '../components';
+import type { DataPath } from '../components';
+import { DataType } from '../data-type';
+import { EffectsObjectData } from '../effects-object-data';
 
 export interface TimelineAssetData extends EffectsObjectData {
   dataType: DataType.TimelineAsset,
